Creates an image ready to instasll OutSystems
TL;DR
-----

Prepares an OVA for use as an OutSystems server

Details
-------

Uses `packer` to build a VM from a Windows 2022 Core ISO to use
as an OutSystems paltform controller or front end. The script
`create-image` creeats the image and converts it to an OVA.

The resulting OVA file has an up-to-date Windows installation, the
Powershell modules to automated an OutSystems setup, and all of the
pre-requisites that OS requires.