Automates image creation with Concourse (#4)
TL;DR
-----

Creates the image with a Concourse pipeline than a script

Details
-------

Implements a Concourse pipeline to automate image creation. The pipeline
uses packer to create an OVA which it then publishes to Minio. The OVA
is ready to go as an OutSystems server regardless of role: it can be a
front end, a platform server, or a LifeTime server.

The pipeline uses vault for secret storage, and the script `populate-vault` 
populates the necessary secrets from `params.yaml`. For non-secret 
parameters, another script (`prepare-params`) creates a parameter file
from `params.yaml` and places it in `${WORK_DIR}/params.yaml`.