Enables WinRM for Ansible
TL;DR
-----

Adds WinRM setup for later Ansible configuration

Details
-------

Steals from the Ansible example[^1] to configure the image so that
Ansible can be used on the machines built from it. The example
approach needs refinement for the aactual machines, but is a good
baseline for the image. Machine customization can harden it up.

[^1]:
https://raw.githubusercontent.com/ansible/ansible/devel/examples/scripts/ConfigureRemotingForAnsible.ps1